requirement.txt
streamlit 
python-dotenv
pypdf
google-generativeai

